+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.2
+
+- [Options.l33tMaxSubstitutions](https://pub.dev/documentation/zxcvbnm/latest/options/Options/l33tMaxSubstitutions.html) now defaults to 100 (was 512).
+- No more than 3 consecutive substitution of the same l33t replacement are made.
+
 ## 1.0.1
 
 - Move translations for Flutter to a separate package.